requirements.txt
# File Converter - Dependencies
Django>=4.2,<5.0
Pillow>=10.0.0
pydub>=0.25.1
pdf2docx>=0.5.8
python-docx>=1.1.0
reportlab>=4.0.0
xhtml2pdf>=0.2.13
gunicorn
dj-database-url>=2.1.0
whitenoise>=6.6.0
# Note: Для аудио/видео конвертации нужен ffmpeg в системе
# В Dockerfile ffmpeg уже установлен
